refactor(adminv2): changes after review
